K-9 refactoring (transfer some requests to RequestMessage) & javadoc

+++ src/main/java/com/edu/mtdr/prpayment/controller/ParticipantController.java
@@ -1,8 +1,8 @@
 package com.edu.mtdr.prpayment.controller;
 
-import com.edu.mtdr.prpayment.config.datasource.DbContextHolder;
-import com.edu.mtdr.prpayment.config.datasource.DbTypeEnum;
 import com.edu.mtdr.prpayment.model.BaseResponseMessage;
+import com.edu.mtdr.prpayment.model.FailureResponseMessage;
+import com.edu.mtdr.prpayment.model.RequestMessage;
 import com.edu.mtdr.prpayment.model.SuccessResponseMessage;
 import com.edu.mtdr.prpayment.repository.ParticipantRepository;
 import com.edu.mtdr.prpayment.schema.ParticipantEntity;
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 
 /**
@@ -33,6 +32,9 @@ public class ParticipantController {
         this.participantRepository = participantRepository;
     }
 
+    /**
+     * @return {@link SuccessResponseMessage} with list of all {@link ParticipantEntity}
+     */
     @GetMapping("/list")
     @ApiOperation("List participants")
     public BaseResponseMessage<?> listParticipants() {
@@ -40,6 +42,10 @@ public class ParticipantController {
         return new SuccessResponseMessage<>(dbParticipants);
     }
 
+    /**
+     * @param id {@link ParticipantEntity}'s id
+     * @return {@link SuccessResponseMessage} with {@link ParticipantEntity}
+     */
     @GetMapping("/get/{id}")
     @ApiOperation("Get participant by id")
     public BaseResponseMessage<ParticipantEntity> getParticipant(@PathVariable("id") Long id) {
@@ -47,16 +53,42 @@ public class ParticipantController {
         return new SuccessResponseMessage<>(dbParticipant);
     }
 
+    /**
+     * @param participant participant to save
+     * @return saved in same state at all shards participant
+     */
     @PostMapping("/save")
     @ApiOperation("Create or update participant")
     public BaseResponseMessage<?> saveParticipant(@RequestBody ParticipantEntity participant) {
         return new SuccessResponseMessage<>(participantService.save(participant));
     }
 
+    /**
+     * @param message {@link RequestMessage} with {@link ParticipantEntity#getId()}
+     *                of {@link ParticipantEntity} to delete
+     * @return {@link SuccessResponseMessage} empty message if all ok, {@link FailureResponseMessage} if exception
+     */
     @PostMapping("/delete")
     @ApiOperation("Delete participant")
-    public BaseResponseMessage<?> deleteParticipant(@RequestBody ParticipantEntity participant) {
-        participantRepository.deleteById(participant.getId());
+    public BaseResponseMessage<?> deleteParticipantById(@RequestBody RequestMessage<Long> message) {
+        participantRepository.deleteById(message.getData());
         return new SuccessResponseMessage<>();
+    }
+
+    /**
+     * @param message {@link RequestMessage} with {@link ParticipantEntity#getName()}
+     *                of {@link ParticipantEntity} to delete
+     * @return {@link SuccessResponseMessage} empty message if all ok, {@link FailureResponseMessage} if exception
+     */
+    @PostMapping("/delete")
+    @ApiOperation("Delete participant")
+    public BaseResponseMessage<?> deleteParticipantByName(@RequestBody RequestMessage<String> message) {
+        ParticipantEntity participant = participantRepository.findFirstByName(message.getData()).orElse(null);
+        if (participant != null) {
+            participantRepository.deleteById(participant.getId());
+            return new SuccessResponseMessage<>();
+        } else {
+            return new FailureResponseMessage<>("Participant not found");
+        }
     }
 }
